fix: correção do README.md

+++ sql/init.sql
@@ -7,7 +7,7 @@ CREATE TABLE usuarios (
 );
 
 CREATE TABLE veiculos (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     placa VARCHAR(10) UNIQUE NOT NULL,
     modelo VARCHAR(50) NOT NULL,
     tipo VARCHAR(20) CHECK (tipo IN ('LEVE', 'PESADO')), 
@@ -15,8 +15,8 @@ CREATE TABLE veiculos (
 );
 
 CREATE TABLE viagens (
-    id SERIAL PRIMARY KEY,
-    veiculo_id INTEGER REFERENCES veiculos(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    veiculo_id BIGINT REFERENCES veiculos(id) ON DELETE CASCADE,
     data_saida TIMESTAMP NOT NULL,
     data_chegada TIMESTAMP,
     origem VARCHAR(100),
@@ -25,8 +25,8 @@ CREATE TABLE viagens (
 );
 
 CREATE TABLE manutencoes (
-    id SERIAL PRIMARY KEY,
-    veiculo_id INTEGER REFERENCES veiculos(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    veiculo_id BIGINT REFERENCES veiculos(id) ON DELETE CASCADE,
     data_inicio DATE NOT NULL,
     data_finalizacao DATE,
     tipo_servico VARCHAR(100),
